remove plugin lists from file that will be come the legacy plugin view
