+++ src/modules/UI/scenes/Plugins/index.js
@@ -1,5 +1,6 @@
 // @flow
 
-import { PluginBuySell, PluginSpend, PluginViewConnect, renderPluginBackButton } from '../../../../components/scenes/PluginViewScene'
+import { PluginBuySell, PluginSpend } from '../../../../components/scenes/PluginViewListScene'
+import { PluginViewConnect, renderPluginBackButton } from '../../../../components/scenes/PluginViewScene'
 
 export { PluginViewConnect as PluginView, PluginBuySell, PluginSpend, renderPluginBackButton }
